requirements.txt
streamlit>=1.28.0
geopy>=2.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
groq>=0.3.0
requests>=2.31.0